Keyguard: Show the SIM unlock dialog once each try
This commit makes the progress dialog's show() get called _only_ once
each time the SIM unlock process is started.

It was reported that a phone can freeze when the PIN code for the SIM
card is entered "too quickly". The user is left with a device on which
the screen is occupied by a non-cancellable dialog box saying that the
SIM card is being unlocked even though it was already unlocked
successfully as it is indicated by the logs. The cause appeared to be
that the dialog box gets shown again after it is goes hidden and
initial reports seem to agree with this as the change introduced by
this commit fixed this issue on the tested devices.

Change-Id: I1c419566219aa98a5a0d6539e9f40cea67b480f5